#7 Add module(sequelize, User) by CommonJS
