+++ sequelize.js
@@ -1,0 +1,13 @@
+var Sequelize = require('sequelize');
+var sequelize = new Sequelize('sqlite:sample.db');
+
+sequelize
+  .authenticate()
+  .then(() => {
+    console.log('Connection has been established successfully.');
+  })
+  .catch(err => {
+    console.error('Unable to connect to the database:', err);
+  });
+
+module.exports = sequelize;
